c: silence the mechanical pedantic warnings
Unused callback parameters are void-cast at the top of the body, per
existing precedent. Dead locals go away, including krun's
computed-but-never-printed receive percentage. The DEFINE_ACQ_IMPL
invocations drop their stray trailing semicolons. kfind's parameters
that shadowed file-scope variables are renamed. sprintf becomes
snprintf where the buffer bound is known. The remaining MSVC CRT
deprecation warnings -- strerror and fopen in single-threaded,
immediate-use call sites where the _s variants would add only
non-portable branching -- are suppressed per target with
_CRT_SECURE_NO_WARNINGS, and the C++ examples get /EHsc. The Windows
UDP layer's sendto length is cast like the file's recvfrom calls.

+++ c/src/udp_win.c
@@ -272,7 +272,7 @@ katherine_udp_send_exact(katherine_udp_t *u, const void *data, size_t count)
     const char *cdata = (const char *) data;
 
     do {
-        sent = sendto(u->sock, cdata + total, count - total, 0, (struct sockaddr *) &u->addr_remote, sizeof(u->addr_remote));
+        sent = sendto(u->sock, cdata + total, (int) (count - total), 0, (struct sockaddr *) &u->addr_remote, sizeof(u->addr_remote));
         if (sent == SOCKET_ERROR) {
             return WSAGetLastError();
         }
